make sendWidget standalone component
