+++ src/views/Dashboard/components/Widgets/SendWidget/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SendWidget";
